// src/application/cuePatternParams/core/0-interface.ts
export interface ICuePatternParams {
  hasWoTopic: boolean;
  hasNiTopic: boolean;
  hasNoneTopic: boolean;
  hasWoGroping: boolean;
  hasNiGroping: boolean;
  hasNoneGroping: boolean;
  hasStraightOrder: boolean;
  hasInvertOrder: boolean;
  hasPositive: boolean;
  hasNegative: boolean;
  hasGroupingTopic: boolean;
}